// src/payload-types.ts
/* tslint:disable */
/* eslint-disable */
/**
 * This file was automatically generated by Payload.
 * DO NOT MODIFY IT BY HAND. Instead, modify your source Payload config,
 * and re-run `payload generate:types` to regenerate this file.
 */

export type CardSlider = {
  title?: string;
  link?: string;
  id?: string;
}[];

export interface Config {
  collections: {
    'contact-requests': ContactRequest;
    events: Event;
    groups: Group;
    users: User;
  };
  globals: {};
}
export interface ContactRequest {
  id: string;
  message: string;
  updatedAt: string;
  createdAt: string;
}
export interface Event {
  id: string;
  title: string;
  start: string;
  end: string;
  description: string;
  organizer?: string | User;
  url: string;
  location: string;
  group: string | Group;
  updatedAt: string;
  createdAt: string;
  _status?: 'draft' | 'published';
}
export interface User {
  id: string;
  firstName: string;
  lastName: string;
  roles?: ('admin' | 'editor')[];
  groups?: string[] | Group[];
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string;
  resetPasswordExpiration?: string;
  salt?: string;
  hash?: string;
  loginAttempts?: number;
  lockUntil?: string;
  password?: string;
}
export interface Group {
  id: string;
  title: string;
  slug: string;
  description?: {
    [k: string]: unknown;
  }[];
  slider?: CardSlider;
  updatedAt: string;
  createdAt: string;
}